+++ src/middlewares/multerProduct.js
@@ -1,0 +1,18 @@
+const path = require("path");
+const multer = require("multer");
+
+const storage = multer.diskStorage({
+  destination: path.join(__dirname, "../../public/img/products"),
+  filename: function (req, file, cb) {
+    cb(
+      null,
+      file.fieldname + "-" + Date.now() + path.extname(file.originalname)
+    );
+  },
+});
+
+const uploadImgProduct = multer({
+  storage: storage,
+});
+
+module.exports = uploadImgProduct;
